requirements.txt
# Western Air Compliance - Desktop PDF Application Requirements
# Install with: pip install -r requirements.txt

# PDF Generation and Manipulation
reportlab>=4.0.0
pypdf>=4.0.0

# Word template support
python-docx>=1.1.0
comtypes>=1.4.0; platform_system == "Windows"

# Image handling (for ReportLab)
Pillow>=10.0.0
